pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chaos-auditor"
version = "1.0.0"
description = "RL environment for training AI agents to find vulnerabilities in self-healing systems"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
include = ["chaos_auditor*"]
